inference/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
gradio>=4.36.0
huggingface_hub>=0.23.0
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.16
safetensors>=0.4.3
Pillow>=10.3.0
opencv-python-headless>=4.9.0
albumentations>=1.4.0
numpy>=1.26.0
python-multipart>=0.0.9
scipy>=1.12.0
scikit-learn>=1.4.0
